@@ -22,12 +22,18 @@
     "webpack-cli": "^4.9.1"
   },
   "dependencies": {
+    "@emotion/is-prop-valid": "^1.3.1",
+    "@heroicons/react": "^2.1.5",
+    "@tanstack/react-query": "^5.59.20",
     "chart.js": "^4.4.3",
     "css-loader": "^7.1.2",
+    "framer-motion": "^11.11.11",
     "postcss-loader": "^8.1.1",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-chartjs-2": "^5.2.0",
-    "react-dom": "^17.0.2",
-    "style-loader": "^4.0.0"
+    "react-dom": "^18.3.1",
+    "react-query": "^3.39.3",
+    "style-loader": "^4.0.0",
+    "zustand": "^5.0.1"
   }
 }
