@@ -1,5 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import App from './App';
 import "./index.css";
-ReactDOM.render(<App />, document.getElementById('root-ext'));
+import { createRoot } from "react-dom/client";
+
+createRoot(document.getElementById("root-ext")!).render(<App />);
