@@ -15,7 +15,7 @@ module.exports = {
       {
         test: /\.tsx?$/,
         use: 'ts-loader',
-        exclude: /node_modules/
+        exclude: [/node_modules/, /server/]
       },
       {
         test: /\.css$/i,
